requirements.txt
fastapi
python-dotenv
pydantic
groq
uvicorn
Flask
